GreeterScreen 분리 및 호이스팅 작업
